Prepare to refactoring core of items

+++ Assets/Project/Scripts/Systems/Item/Base/BaseData.cs
@@ -5,7 +5,7 @@ public class BaseData
 {
     public string name;
     public string description;
-    public float cost;
+    public int cost;
     public int id;
     public bool isBlocked;
     public Sprite icon;
